Added the modal for the add grader for code
