+++ devU-client/src/components/pages/forms/assignments/assignmentUpdatePage.tsx
@@ -18,6 +18,9 @@ import { Button as MuiButton, StyledEngineProvider } from '@mui/material'
 import TextProblemModal from './textProblemModal'
 /*import Dropdown, { Option } from 'components/shared/inputs/dropdown';*/
 //import Select from 'react-select/src/Select'
+import ContainerAutoGraderModal from '../containers/ContainerAutoGraderModal';
+import AddProblemModal from './AddProblemModal'
+
 
 type UrlParams = { assignmentId: string }
 
@@ -28,7 +31,13 @@ const AssignmentUpdatePage = () => {
   const currentAssignmentId = parseInt(assignmentId)
   const [assignmentProblems, setAssignmentProblems] = useState<AssignmentProblem[]>([])
   const [nonContainerAutograders, setNonContainerAutograders] = useState<NonContainerAutoGrader[]>([])
+  const [containerAutoGraderModal, setContainerAutoGraderModal] = useState(false);
+  const handleCloseContainerAutoGraderModal = () => setContainerAutoGraderModal(false);
   const [containerAutograders, setContainerAutograders] = useState<ContainerAutoGrader[]>([])
+  const [addProblemModal, setAddProblemModal] = useState(false);
+  const handleCloseAddProblemModal = () => setAddProblemModal(false);
+
+  
 
   /*const [categories, setCategories] = useState<Category[]>([])
   const [categoryOptions, setAllCategoryOptions] = useState<Option<Category>[]>([])*/
@@ -236,11 +245,13 @@ const AssignmentUpdatePage = () => {
         fetchAssignmentProblems()
     })
   }
+
   
   return (
+    <>
+      <ContainerAutoGraderModal open={containerAutoGraderModal} onClose={handleCloseContainerAutoGraderModal} />
 
-    <PageWrapper>
-
+      <PageWrapper>
       <Dialog open={openEditModal} onClose={handleCloseModal}>
         <DialogContent sx={{bgcolor:theme.listItemBackground}}>
         <h3 className={styles.header}>Edit Problem</h3>
@@ -254,6 +265,7 @@ const AssignmentUpdatePage = () => {
       </Dialog>
 
       <TextProblemModal open={textModal} onClose={handleCloseTextModal}/>
+      <AddProblemModal open={addProblemModal} onClose={handleCloseAddProblemModal} />
 
       <div className={styles.pageHeader}>
         <h1 style={{gridColumnStart:2}}>Edit Assignment</h1>
@@ -349,21 +361,20 @@ const AssignmentUpdatePage = () => {
             </MuiButton>
           </StyledEngineProvider>
           </label> 
-          
 
         </div>
         <div className={styles.problemsList}>
         <h2 className={styles.header}>Add Problems</h2>
           <div className={styles.buttonContainer}>
-            <Button onClick={() => setAlert({ autoDelete: true, type: 'error', message: 'Setup Code/File Input creation modal' })} className='btnSecondary'>Code/File Input</Button>
+          <Button onClick={() => setAddProblemModal(true)} className='btnSecondary'>Code/File Input</Button>
             <Button onClick={() => {setTextModal(true)}} className='btnSecondary'>Text Input</Button>
             <Button onClick={() => setAlert({ autoDelete: true, type: 'error', message: 'Setup Multiple Choice creation modal' })} className='btnSecondary'>Multiple Choice</Button>
           </div>
           <h2 className={styles.header}>Add Graders</h2>
           <div className={styles.buttonContainer}>
-            <Button onClick={() => {
-                        history.push(`createCAG`)
-                    }} className='btnSecondary'>Code Grader</Button>
+          <Button onClick={() => setContainerAutoGraderModal(true)} className='btnSecondary'>
+            Code Grader
+          </Button>
             <Button onClick={() => {
                         history.push(`createNCAG`)
                     }} className='btnSecondary'>Non-code Grader</Button>
@@ -396,7 +407,8 @@ const AssignmentUpdatePage = () => {
       <div style={{ display: 'flex', justifyContent: 'center', margin: '20px'}}>
             <Button onClick={handleAssignmentUpdate} className='btnPrimary'>save and exit</Button>
       </div>
-    </PageWrapper>
+      </PageWrapper>
+    </>
   )
 }
